refactor(#4229): extract terminal token update tail
