+++ src/services/discord/tmux_watcher.rs
@@ -160,6 +160,9 @@ mod turn_stream_collector;
 #[path = "tmux_watcher/post_stream_exit.rs"]
 mod post_stream_exit;
 
+#[path = "tmux_watcher/terminal_token_update.rs"]
+mod terminal_token_update;
+
 pub(in crate::services::discord) use self::completion_gate::{
     TuiCompletionGateOutcome, run_tui_completion_gate,
 };
@@ -177,6 +180,7 @@ use self::supervisor_relay::*;
 use self::terminal_abort_exits::*;
 use self::terminal_commit_epilogue::*;
 use self::terminal_readiness::*;
+use self::terminal_token_update::*;
 #[cfg(test)]
 pub(in crate::services::discord) use self::turn_identity::pinned_delivery_lease_key as pinned_delivery_lease_key_for_test;
 use self::turn_stream_collector::*;
@@ -3422,134 +3426,17 @@ pub(in crate::services::discord) async fn tmux_output_watcher_with_restore(
             );
         }
 
-        // Update session tokens from result event and auto-compact if threshold exceeded
-        if let Some(tokens) = result_usage.map(|usage| usage.context_occupancy_input_tokens()) {
-            let provider = shared.settings.read().await.provider.clone();
-            let session_key = crate::services::discord::adk_session::build_adk_session_key(
-                &shared, channel_id, &provider, None,
-            )
-            .await;
-            let channel_name = {
-                let data = shared.core.lock().await;
-                data.sessions
-                    .get(&channel_id)
-                    .and_then(|s| s.channel_name.clone())
-            };
-            let thread_channel_id = channel_name
-                .as_deref()
-                .and_then(crate::services::discord::adk_session::parse_thread_channel_id_from_name);
-            let agent_id = resolve_role_binding(channel_id, channel_name.as_deref())
-                .map(|binding| binding.role_id);
-            crate::services::discord::adk_session::post_adk_session_status(
-                session_key.as_deref(),
-                channel_name.as_deref(),
-                None,
-                watcher_terminal_token_update_status(watcher_direct_terminal_idle_committed),
-                &provider,
-                None,
-                Some(tokens),
-                None,
-                None,
-                thread_channel_id,
-                Some(channel_id),
-                agent_id.as_deref(),
-                shared.api_port,
-            )
-            .await;
-
-            let ctx_cfg =
-                crate::services::discord::adk_session::fetch_context_thresholds(shared.api_port)
-                    .await;
-            let pct = (tokens * 100) / ctx_cfg.context_window.max(1);
-            // #227: Re-enabled with 5-min cooldown (matches turn_bridge path).
-            // Without cooldown, the compact turn's own result could re-trigger compact.
-            let cooldown_key = format!("auto_compact_cooldown:{}", channel_id.get());
-            let cooldown_value =
-                match crate::services::discord::internal_api::get_kv_value(&cooldown_key) {
-                    Ok(value) => value,
-                    Err(_) => {
-                        if let Some(pg_pool) = shared.pg_pool.as_ref() {
-                            sqlx::query_scalar::<_, Option<String>>(
-                                "SELECT value
-                             FROM kv_meta
-                             WHERE key = $1
-                               AND (expires_at IS NULL OR expires_at > NOW())
-                             LIMIT 1",
-                            )
-                            .bind(&cooldown_key)
-                            .fetch_optional(pg_pool)
-                            .await
-                            .ok()
-                            .flatten()
-                            .flatten()
-                        } else {
-                            None
-                        }
-                    }
-                };
-            let compact_cooldown_ok =
-                cooldown_value
-                    .and_then(|v| v.parse::<i64>().ok())
-                    .map_or(true, |ts| {
-                        let now = std::time::SystemTime::now()
-                            .duration_since(std::time::UNIX_EPOCH)
-                            .unwrap_or_default()
-                            .as_secs() as i64;
-                        now - ts > 300 // 5 min cooldown
-                    });
-            // DISABLED — token counting still unreliable
-            if false && pct >= ctx_cfg.compact_pct && !is_prompt_too_long && compact_cooldown_ok {
-                let ts = chrono::Local::now().format("%H:%M:%S");
-                tracing::warn!(
-                    "  [{ts}] ⚡ [watcher] Auto-compact: {} at {pct}% ({tokens} tokens)",
-                    tmux_session_name
-                );
-                let name = tmux_session_name.clone();
-                let _ = tokio::task::spawn_blocking(move || {
-                    crate::services::platform::tmux::send_keys(&name, &["/compact", "Enter"])
-                })
-                .await;
-                // Set cooldown timestamp
-                let cooldown_key = format!("auto_compact_cooldown:{}", channel_id.get());
-                let now = std::time::SystemTime::now()
-                    .duration_since(std::time::UNIX_EPOCH)
-                    .unwrap_or_default()
-                    .as_secs();
-                let now_text = now.to_string();
-                if crate::services::discord::internal_api::set_kv_value(&cooldown_key, &now_text)
-                    .is_err()
-                {
-                    if let Some(pg_pool) = shared.pg_pool.as_ref() {
-                        let _ = sqlx::query(
-                            "INSERT INTO kv_meta (key, value, expires_at)
-                             VALUES ($1, $2, NULL)
-                             ON CONFLICT (key) DO UPDATE
-                             SET value = EXCLUDED.value,
-                                 expires_at = EXCLUDED.expires_at",
-                        )
-                        .bind(&cooldown_key)
-                        .bind(&now_text)
-                        .execute(pg_pool)
-                        .await;
-                    }
-                }
-                // Notify: auto-compact triggered
-                let target = format!("channel:{}", channel_id.get());
-                let content = format!("🗜️ 자동 컨텍스트 압축 (사용률: {pct}%)");
-                let _ = enqueue_outbox_best_effort(
-                    shared.pg_pool.as_ref(),
-                    OutboxMessage {
-                        target: target.as_str(),
-                        content: content.as_str(),
-                        bot: "notify",
-                        source: "system",
-                        reason_code: None,
-                        session_key: None,
-                    },
-                )
-                .await;
-            }
-        }
+        // #4229 W7b S-A: terminal token update tail moved verbatim to
+        // tmux_watcher/terminal_token_update.rs.
+        run_watcher_terminal_token_update(WatcherTerminalTokenUpdateContext {
+            result_usage,
+            shared: &shared,
+            channel_id,
+            watcher_direct_terminal_idle_committed,
+            tmux_session_name: &tmux_session_name,
+            is_prompt_too_long,
+        })
+        .await;
     }
 
     // #4229 S5: post-stream-exit finalize tail moved verbatim to tmux_watcher/post_stream_exit.rs.
